docs: add documentation for plugin Sync
