+++ docs/package.json
@@ -17,6 +17,7 @@
     "@docusaurus/core": "^2.0.0-alpha.73",
     "@docusaurus/preset-classic": "^2.0.0-alpha.73",
     "@docusaurus/theme-live-codeblock": "^2.0.0-alpha.73",
+    "@egjs/flicking-plugins": "^4.2.0",
     "@egjs/react-flicking": "^4.2.1",
     "@egjs/react-grid": "^1.1.2",
     "@mdx-js/react": "^1.6.21",
@@ -42,7 +43,6 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-decorators": "^7.13.15",
-    "@egjs/flicking-plugins": "^4.1.0-beta.1",
     "remark-breaks": "^2.0.2"
   }
 }
