+++ docs/src/pages/Demos.mdx
@@ -9,7 +9,6 @@ import AddRemove from "@site/src/demo/AddRemove";
 import SideScroll from "@site/src/demo/SideScroll";
 import Grid from "@site/src/demo/Grid";
 import WeatherDisplay from "@site/src/demo/WeatherDisplay";
-import Sync from "@site/src/demo/Sync";
 
 # Demos
 ## Panel add / remove
@@ -42,7 +41,3 @@ Made with [@egjs/grid](https://github.com/naver/egjs-grid)
 
 ## Vertical Display
 <WeatherDisplay />
-
-## Sync between Flickings
-You can create Clubhouse Topics/Interests-like UI with Flicking
-<Sync />
