+++ docs/src/pages/Plugins.mdx
@@ -11,6 +11,7 @@ import ParallaxDemo from "@site/src/demo/plugins/Parallax";
 import FadeDemo from "@site/src/demo/plugins/Fade";
 import ArrowDemo from "@site/src/demo/plugins/Arrow";
 import PaginationDemo from "@site/src/demo/plugins/Pagination";
+import SyncDemo from "@site/src/demo/plugins/Sync";
 import "@site/src/css/demo/plugins.css";
 
 # Plugins
@@ -242,3 +243,6 @@ These are the available CSS files:
 |fractionCurrentFormat|function|✔️|(index: number) => index.toString()|A format to display a current index inside `fraction-current` class|
 |fractionTotalFormat|function|✔️|(index: number) => index.toString()|A format to display a total panels count inside `fraction-total` class|
 |scrollOnChange|function|✔️|(index: number, ctx: ScrollContext) => ctx.moveTo(index)|A callback function triggered when the active bullet changes|
+
+## Sync
+<SyncDemo />
