+++ packages/ngx-flicking/package.json
@@ -31,7 +31,7 @@
     "@angular/compiler-cli": "~11.2.4",
     "@angular/language-service": "~11.2.4",
     "@egjs/flicking": "^4.1.0",
-    "@egjs/flicking-plugins": "^4.1.0-beta.1",
+    "@egjs/flicking-plugins": "^4.2.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
