+++ packages/ngx-flicking/projects/demo/src/app/app.component.html
@@ -14,5 +14,6 @@
   <a routerLink="/custom" routerLinkActive="active">Custom</a>
   <a routerLink="/visiblepanels" routerLinkActive="active">Render Only Visible</a>
   <a routerLink="/arrow" routerLinkActive="active">Arrow</a>
+  <a routerLink="/sync" routerLinkActive="active">Sync</a>
 </nav>
 <router-outlet></router-outlet>
