+++ packages/ngx-flicking/projects/demo/src/app/sync/sync.component.html
@@ -1,0 +1,20 @@
+<div id="sync" class="container">
+  <ngx-flicking class="flicking" #flicking0 [options]="{ bound: true, bounce: 30, align: 'prev' }" [plugins]="plugins">
+    <span flicking-panel class="panel button mr-2 is-white">🍎 Apple</span>
+    <span flicking-panel class="panel button mr-2 is-white">🍉 Watermelon</span>
+    <span flicking-panel class="panel button mr-2 is-white">🥝 Kiwi</span>
+    <span flicking-panel class="panel button mr-2 is-white">...</span>
+  </ngx-flicking>
+  <ngx-flicking class="flicking" #flicking1 [options]="{ bound: true, bounce: 30, align: 'prev' }">
+    <span flicking-panel class="panel button mr-2 is-white">🍔 Hamburger</span>
+    <span flicking-panel class="panel button mr-2 is-white">🍕 Pizza</span>
+    <span flicking-panel class="panel button mr-2 is-white">🍞 Bread</span>
+    <span flicking-panel class="panel button mr-2 is-white">...</span>
+  </ngx-flicking>
+  <ngx-flicking class="flicking" #flicking2 [options]="{ bound: true, bounce: 30, align: 'prev' }">
+    <span flicking-panel class="panel button mr-2 is-white">🥛 Milk</span>
+    <span flicking-panel class="panel button mr-2 is-white">☕ Coffee</span>
+    <span flicking-panel class="panel button mr-2 is-white">🍵 Green tea</span>
+    <span flicking-panel class="panel button mr-2 is-white">...</span>
+  </ngx-flicking>
+</div>
