+++ packages/vue-flicking/demo/Demo.vue
@@ -18,6 +18,7 @@
         <router-link to="/fade">Fade</router-link>
         <router-link to="/parallax">Parallax</router-link>
         <router-link to="/arrow">Arrow</router-link>
+        <router-link to="/sync">Sync</router-link>
       </div>
     </header>
     <transition name="component-fade" mode="out-in">
