+++ packages/vue-flicking/package.json
@@ -42,7 +42,7 @@
   },
   "devDependencies": {
     "@egjs/build-helper": "0.0.5",
-    "@egjs/flicking-plugins": "^4.1.0-beta.1",
+    "@egjs/flicking-plugins": "^4.2.0",
     "@types/chai": "^4.1.0",
     "@types/mocha": "^5.2.4",
     "@types/uuid": "^8.3.0",
